Day 20 | Fixed Failing Unit Tests (AdventMap Refactoring Post-Merge)

+++ implementation/src/main/kotlin/com/aoc/intcode/tractorbeam/ShipNotFound.kt
@@ -2,4 +2,4 @@ package com.aoc.intcode.tractorbeam
 
 import java.lang.IllegalArgumentException
 
-class ShipNotFound(areaSize: Int) : IllegalArgumentException("Could not find beam square of size $areaSize")
+class ShipNotFound(areaSize: Long) : IllegalArgumentException("Could not find beam square of size $areaSize")
